move errors from api v1, add RateLimitExceededError #55
