+++ api/errors.js
@@ -121,6 +121,22 @@ function ServerError(origin, resource, message) {
 }
 ServerError.prototype = UbicallError.prototype;
 
+/**
+ * return HTTP 429 error with @param message or  "Rate limit exceeded"
+ * @param origin - original error object
+ * @param resource - current resource url path
+ * @param message - custom error message
+ * @return HTTP 429
+ * @memberof Error
+ */
+function RateLimitExceededError(origin, resource, message) {
+    this.name = "RateLimitExceeded Error";
+    this.status = 429;
+    this.response = {};
+    this.response.resource = resource;
+    this.response.message = message || "Rate limit exceeded";
+}
+RateLimitExceededError.prototype = UbicallError.prototype;
 
 module.exports = {
     UbicallError: UbicallError,
@@ -129,5 +145,6 @@ module.exports = {
     MissedParams: MissedParams,
     Forbidden: Forbidden,
     ServerError: ServerError,
-    NotFound: NotFound
+    NotFound: NotFound,
+    RateLimitExceededError: RateLimitExceededError
 };
