+++ api/v1/3rd/index.js
@@ -13,7 +13,7 @@ var multer = require("multer");
 var log = require("../../../log");
 var zendesk = require("./zendesk/ticket");
 var midware = require("../utils/midware");
-var errorHandler = require("../utils/errorHandler");
+var errorHandler = require("../../errorHandler");
 var needsPermission = require("ubicall-oauth").needsPermission;
 var hasZendeskCredinitial = require("./zendesk/utils.js").hasZendeskCredinitial;
 var mzdmr = require("./zendesk/utils.js").matchZendeskMinimumTicketRequirement;
