+++ api/v1/call.js
@@ -12,12 +12,12 @@ var settings = require("../../settings");
 var storage = require("../../storage");
 var log = require("../../log");
 var infra = require("../../infra");
-var NotImplementedError = require("./utils/errors").NotImplementedError;
-var BadRequest = require("./utils/errors").BadRequest;
-var MissedParams = require("./utils/errors").MissedParams;
-var Forbidden = require("./utils/errors").Forbidden;
-var ServerError = require("./utils/errors").ServerError;
-var NotFound = require("./utils/errors").NotFound;
+var NotImplementedError = require("../errors").NotImplementedError;
+var BadRequest = require("../errors").BadRequest;
+var MissedParams = require("../errors").MissedParams;
+var Forbidden = require("../errors").Forbidden;
+var ServerError = require("../errors").ServerError;
+var NotFound = require("../errors").NotFound;
 
 /**
  * helper function used to sumbit demo call by calling storage.scheduleDemoCall
