+++ api/v1/index.js
@@ -20,13 +20,13 @@ var queue = require("./queue");
 var ivr = require("./ivr");
 var zendesk = require("./3rd/zendesk/ticket");
 var midware = require("./utils/midware");
-var errorHandler = require("./utils/errorHandler");
-var NotImplementedError = require("./utils/errors").NotImplementedError;
-var BadRequest = require("./utils/errors").BadRequest;
-var MissedParams = require("./utils/errors").MissedParams;
-var Forbidden = require("./utils/errors").Forbidden;
-var ServerError = require("./utils/errors").ServerError;
-var NotFound = require("./utils/errors").NotFound;
+var errorHandler = require("../errorHandler");
+var NotImplementedError = require("../errors").NotImplementedError;
+var BadRequest = require("../errors").BadRequest;
+var MissedParams = require("../errors").MissedParams;
+var Forbidden = require("../errors").Forbidden;
+var ServerError = require("../errors").ServerError;
+var NotFound = require("../errors").NotFound;
 var needsPermission = require("ubicall-oauth").needsPermission;
 var passport = require("passport");
 var helmet = require("helmet");
