+++ api/v1/queue.js
@@ -10,12 +10,12 @@ var request = require("request");
 var settings = require("../../settings");
 var storage = require("../../storage");
 var log = require("../../log");
-var NotImplementedError = require("./utils/errors").NotImplementedError;
-var BadRequest = require("./utils/errors").BadRequest;
-var MissedParams = require("./utils/errors").MissedParams;
-var Forbidden = require("./utils/errors").Forbidden;
-var ServerError = require("./utils/errors").ServerError;
-var NotFound = require("./utils/errors").NotFound;
+var NotImplementedError = require("../errors").NotImplementedError;
+var BadRequest = require("../errors").BadRequest;
+var MissedParams = require("../errors").MissedParams;
+var Forbidden = require("../errors").Forbidden;
+var ServerError = require("../errors").ServerError;
+var NotFound = require("../errors").NotFound;
 /**
  * @return HTTP status - 200
  * @return HTTP status - 404 {@link NotFound} If can"t return queue data
